requirements.txt
# Snowpipe Streaming High-Performance Architecture Requirements
# MTA Real-time Transit Data Pipeline

# Core Snowflake dependencies
snowflake-snowpark-python==1.11.1
snowflake-connector-python==3.15.0
snowflake-ingest==1.0.11

# Cryptographic libraries (for key-pair authentication)
cryptography>=41.0.7,<46.0.0

# Data processing and transformation (NumPy 1.x compatible)
pandas>=2.1.4,<2.2.0  # Latest stable pandas with NumPy 1.x support
numpy>=1.24.4,<2.0.0  # Explicitly prevent NumPy 2.x compatibility issues
pyarrow>=14.0.2,<15.0.0  # Compatible with pandas 2.1.x

# Real-time streaming and messaging (PRIMARY)
confluent-kafka==2.3.0  # Kafka consumer with SASL_SSL support
pika==1.3.2             # RabbitMQ (optional)
redis==5.0.1            # Caching and coordination

# HTTP and API clients
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

# Data validation and quality
jsonschema==4.20.0
cerberus==1.3.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For enhanced performance
uvloop==0.19.0  # Linux/macOS only
cython==3.0.6